Created page citiesAdmin that allows to get all the created cities from dataBase and delete the selected city. changed findAll in cities controller to obtain the related province information

+++ backend/src/cities/cities.controller.ts
@@ -22,7 +22,7 @@ const em = orm.em
 
 async function findAll(req: Request, res: Response) {
   try {
-    const cities = await em.find(City, {})
+    const cities = await em.find(City, {}, {populate: ['province']})
     res.status(200).json({ message: 'find all cities', data: cities })
   }catch (error : any) {
     res.status(500).json({ message: error.message })
@@ -66,6 +66,9 @@ async function update(req: Request, res: Response) {
 async function remove(req: Request, res: Response) {
   try{
     const id = Number.parseInt(req.params.idCity)
+    if (isNaN(id)) {
+      return res.status(400).json({ message: "Invalid city id" })
+    }
     const city = em.getReference(City, id as any)
     await em.removeAndFlush(city)
     res.status(200).json({ message: 'City removed successfully' })
